Quick fix because I'm bad at C

+++ tests/test-lab.c
@@ -212,7 +212,7 @@ void test_buddy_malloc_exhaustion(void)
   void *c = buddy_malloc(&pool, 1);
   void *d = buddy_malloc(&pool, 1);
 
-  for (int i = SMALLEST_K; i <= pool.kval_m; i++) {
+  for (size_t i = SMALLEST_K; i <= pool.kval_m; i++) {
     printf("K: %d, Tag code: %d", i, pool.avail[i].tag);
   }
